+++ workalendar/registry.py
@@ -1,5 +1,4 @@
 from importlib import import_module
-import warnings
 
 from .core import Calendar
 from .exceptions import ISORegistryError
@@ -90,24 +89,6 @@ class IsoRegistry:
             if key.startswith("{}-".format(iso_code)):
                 items[key] = value
         return items
-
-    def items(self, region_codes=None, include_subregions=False):
-        """
-        Returns calendar classes for regions
-
-        :param region_codes list of ISO codes for selected regions. If empty,
-                            the function will return all items from the
-                            registry.
-        :param include_subregions boolean if subregions
-        of selected regions should be included in result
-        :rtype dict
-        :return dict where keys are ISO codes strings
-        and values are calendar classes
-        """
-        warnings.warn("The ``items()`` method will soon be deprecated."
-                      " Please use ``get_calendars()`` instead.",
-                      DeprecationWarning)
-        return self.get_calendars(region_codes, include_subregions)
 
     def get_calendars(self, region_codes=None, include_subregions=False):
         """